coupon/api/src/test/java/spring/coupon/service/ApplyServiceTest.java
package spring.coupon.service;

import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import spring.coupon.repository.CouponRepository;

import static org.assertj.core.api.Assertions.*;
import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest
class ApplyServiceTest {

    @Autowired
    private ApplyService applyService;

    @Autowired
    private CouponRepository couponRepository;

    @Test
    public void 한번만응모() {
        //given

        //when
        applyService.apply(1L);
        long count = couponRepository.count();

        //then
        assertThat(count).isEqualTo(1L);
    }
}